Fixed manual loading of SPU's by setting an environment variable telling whether or not it was loaded manually or by the faker.
Added "startdir" to the mothership so that apps like Q3 can start in the
right place to find their stupid files

+++ util/tcpip.c
@@ -552,7 +552,7 @@ void crTCPIPDoConnect( CRConnection *conn )
 
 	memset( &servaddr, 0, sizeof(servaddr) );
 	servaddr.sin_family = AF_INET;
-	servaddr.sin_port = htons( conn->port );
+	servaddr.sin_port = htons( (short) conn->port );
 
 	memcpy( (char *) &servaddr.sin_addr, hp->h_addr,
 			sizeof(servaddr.sin_addr) );
